+++ src/macpan2.cpp
@@ -136,8 +136,9 @@ enum macpan2_func
     , MP2_PNORM = 57 // fwrap: pnorm(q, mean, sd)
     , MP2_INVLOGIT = 58 // fwrap: invlogit(x)
     , MP2_LOGIT = 59 // fwrap: logit(x)
-    , MP2_ASSIGN = 60 // fwrap: assign(x, i, j, v)
-    , MP2_UNPACK = 61 // fwrap: unpack(x, ...)
+    , MP2_CUMSUM = 60 // fwrap: cumsum(x)
+    , MP2_ASSIGN = 61 // fwrap: assign(x, i, j, v)
+    , MP2_UNPACK = 62 // fwrap: unpack(x, ...)
 };
 
 enum macpan2_meth
@@ -1960,13 +1961,41 @@ public:
             // #' ```
             // #'
 
+            // #' ## Sweeping Matrix Elements
+            // #'
+            // #' ### Functions
+            // #'
+            // #' * `cumsum(x)` : Return a matrix with columns containing the
+            // #' cumulative sum of the columns in `x`.
+            // #'
+            // #' ### Arguments
+            // #'
+            // #' * `x` : A matrix.
+            // #'
+            // #' ### Return
+            // #' 
+            // #' A matrix the same size as `x` but with columns containing the
+            // #' cumulative sum of the columns in `x`.
+            case MP2_CUMSUM:
+                rows = args.rows(0);
+                cols = args.cols(0);
+                m = matrix<Type>::Zero(rows, cols);
+                m1 = args[0];
+                for (int j = 0; j < cols; j++) {
+                    m.coeffRef(0, j) = m1.coeff(0, j);
+                    for (int i = 1; i < rows; i++) {
+                        m.coeffRef(i, j) = m.coeff(i - 1, j) + m1.coeff(i, j);
+                    }
+                }
+                return m;
+            
             // #' ## Extracting Matrix Elements
             // #'
             // #' ### Functions
             // #'
-            // #' * `x[i,j]` : Matrix containing a subset
+            // #' * `x[i,j]` : Return a matrix containing a subset
             // #' of the rows and columns of `x`.
-            // #' * `block(x,i,j,n,m)` : Matrix containing a
+            // #' * `block(x,i,j,n,m)` : Return a matrix containing a
             // #' contiguous subset of rows and columns of `x`
             // #' \url{https://eigen.tuxfamily.org/dox/group__TutorialBlockOperations.html}.
             // #' * `last(x)` : The last element of a matrix (i.e., the
@@ -2379,6 +2408,33 @@ public:
                     MP2_ERR(MP2_TIME_VAR, "The first element of the second argument must be less than the number of elements in the first.", MP2_TIME_VAR);
                     return m;
                 }
+                
+                // off is the 'initial offset', which lets the user
+                // tells us when (i.e., what time-step) we should start
+                // increment.
+                // 
+                // we have switched this initial offset in the user 
+                // interface from zero-based to one-based, which makes
+                // much more sense given that time steps are one-based 
+                // in macpan2. however we are keeping offsets
+                // zero-based in the code, because indexing is
+                // zero-based and because of the following minor
+                // back-compatibility argument.
+                // 
+                // the most common case was to set initial offset = 0,
+                // meaning we start to increment right away at the
+                // first time step. so for back-compatibility we 
+                // allow this to mean the same thing as it did before. 
+                // in the much less common case of offset > 0
+                // (don't think anyone really did this because
+                // it wasn't documented) we break back-compatability
+                // by shifting user-supplied initial offsets down 
+                // by one, which will not matter much for simulations
+                // with a 'reasonable' number of simulation steps.
+                if (t == 1) {
+                  off = off - 1;
+                  if (off < 0) off = 0; 
+                }
 
                 // first argument can have its rows indexed
                 // by the second (curly braces wrap ints in
@@ -2388,11 +2444,9 @@ public:
                 // used in this way require c++11 i believe.)
                 if (off < v.size() - 1) { // might need to increment
                     cp = v[off + 1];
-                    if (cp == t) {                  // yes we need to increment
+                    if (cp == t) { // yes we need to increment
                         off = off + 1; // so we increment
                         matIndex = index2mats[1];
-                        // FIXME: should really have a function that
-                        // sets matrix or int_vec as appropriate
                         if (index2what[1] == 1) { // int-vec-valued pointer
                             // store the new offset in the zeroth position
                             valid_int_vecs.setNthIntVec(matIndex, off, 0);