+++ tests/testthat/test-dot-layout.R
@@ -1,9 +1,11 @@
 test_that("dot layouts produce appropriate errors and output classes", {
   if (require(Rgraphviz, quietly = TRUE)) {
       specs <- mp_tmb_entire_library()
-      no_mpflows <- c("lotka_volterra_competition",
-                  "lotka_volterra_predator_prey",
-                  "nfds")
+      n_flows = (specs
+        |> lapply(mp_flow_frame, topological_sort = FALSE) 
+        |> vapply(nrow, integer(1L))
+      )
+      no_mpflows <- which(n_flows == 0L) |> names()
       for (s in no_mpflows) {
           expect_error(dot_layout(specs[[s]]), "was spec defined")
       }